+++ sentry-android-core/src/main/java/io/sentry/android/core/ManifestMetadataReader.java
@@ -37,6 +37,7 @@ final class ManifestMetadataReader {
   static final String AUTO_INIT = "io.sentry.auto-init";
   static final String NDK_ENABLE = "io.sentry.ndk.enable";
   static final String NDK_SCOPE_SYNC_ENABLE = "io.sentry.ndk.scope-sync.enable";
+  static final String NDK_SDK_NAME = "io.sentry.ndk.sdk-name";
   static final String RELEASE = "io.sentry.release";
   static final String ENVIRONMENT = "io.sentry.environment";
   static final String SDK_NAME = "io.sentry.sdk.name";
@@ -251,6 +252,12 @@ final class ManifestMetadataReader {
 
         options.setEnableScopeSync(
             readBool(metadata, logger, NDK_SCOPE_SYNC_ENABLE, options.isEnableScopeSync()));
+
+        final @Nullable String nativeSdkName =
+            readString(metadata, logger, NDK_SDK_NAME, options.getNativeSdkName());
+        if (nativeSdkName != null) {
+          options.setNativeSdkName(nativeSdkName);
+        }
 
         options.setRelease(readString(metadata, logger, RELEASE, options.getRelease()));