+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Features
+
+- Added `io.sentry.ndk.sdk-name` Android manifest option to configure the native SDK's name ([#5027](https://github.com/getsentry/sentry-java/pull/5027))
+
+## 8.30.0
+
 ### Fixes
 
 - Fix ANRs when collecting device context ([#4970](https://github.com/getsentry/sentry-java/pull/4970))